Move ColumnAssociationResolver to column_assoc.py.
