+++ metricflow/specs/column_assoc.py
@@ -1,8 +1,12 @@
-from abc import ABC
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import Any
 
 from dbt_semantic_interfaces.dataclass_serialization import SerializableDataclass
+
+from metricflow.specs.specs import InstanceSpec
 
 
 class ColumnCorrelationKey(ABC):
@@ -38,3 +42,25 @@ class ColumnAssociation(SerializableDataclass):
     @property
     def column_correlation_key(self) -> ColumnCorrelationKey:  # noqa: D
         return self.single_column_correlation_key
+
+
+class ColumnAssociationResolver(ABC):
+    """Get the default column associations for an element instance.
+
+    This is used for naming columns in an SQL query consistently. For example, dimensions with links are
+    named like <entity link>__<dimension name> e.g. user_id__country, and time dimensions at a different time
+    granularity are named <time dimension>__<time granularity> e.g. ds__month. Having a central place to name them will
+    make it easier to change this later on. Names generated need to be unique within a query.
+
+    It's also important to maintain this format because customers write constraints in SQL assuming this. This
+    allows us to stick the constraint in as WHERE clauses without having to parse the constraint SQL.
+
+    TODO: Updates are needed for time granularity in time dimensions, ToT for metrics.
+
+    The resolve* methods should return the column associations / column names that it should use in queries for the given
+    spec.
+    """
+
+    @abstractmethod
+    def resolve_spec(self, spec: InstanceSpec) -> ColumnAssociation:  # noqa: D
+        raise NotImplementedError
