+++ metricflow/test/model/test_where_filter_spec.py
@@ -4,12 +4,12 @@ from dbt_semantic_interfaces.objects.filters.where_filter import WhereFilter
 from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.references import EntityReference
 from metricflow.specs.specs import (
-    ColumnAssociationResolver,
     LinkableSpecSet,
     DimensionSpec,
     EntitySpec,
     TimeDimensionSpec,
 )
+from metricflow.specs.column_assoc import ColumnAssociationResolver
 from metricflow.specs.where_filter_transform import ConvertToWhereSpec
 
 logger = logging.getLogger(__name__)
